refactor: move component operations to component store
